requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
